moved User MSI validation to msi package
