+++ azurerm/internal/services/msi/validate/identity.go
